requirements.txt
APScheduler==3.11.0
asgiref==3.9.1
black==25.1.0
certifi==2025.7.14
charset-normalizer==3.4.2
click==8.2.1
Django==5.2.4
django-apscheduler==0.7.0
django-unfold==0.63.0
djangorestframework==3.16.0
gunicorn==23.0.0
idna==3.10
migrate==0.3.8
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
platformdirs==4.3.8
python-dotenv==1.1.1
requests==2.32.4
sqlparse==0.5.3
tzlocal==5.3.1
urllib3==2.5.0
whitenoise==6.9.0
